# loms/api/openapiv2/loms.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "LOMS API",
    "description": "API for LOMS",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "LOMS",
      "description": "LOMS service"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/orders/create": {
      "post": {
        "summary": "Create a new order",
        "description": "Creates a new order with the specified items for a user.",
        "operationId": "LOMS_CreateOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "items",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "$ref": "#/definitions/v1Item"
              }
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/v1/orders/{orderId}": {
      "get": {
        "summary": "Get order information",
        "description": "Retrieves information about a specific order using its ID.",
        "operationId": "LOMS_GetOrderInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetOrderInfoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/v1/orders/{orderId}/cancel": {
      "post": {
        "summary": "Cancel an order",
        "description": "Cancels a specified order.",
        "operationId": "LOMS_CancelOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CancelOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/v1/orders/{orderId}/pay": {
      "post": {
        "summary": "Pay for an order",
        "description": "Processes payment for a specified order.",
        "operationId": "LOMS_PayOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1PayOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/v1/stocks/{sku}": {
      "get": {
        "summary": "Get stock information",
        "description": "Retrieves stock information for a specific SKU.",
        "operationId": "LOMS_GetStockInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetStockInfoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "sku",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": [
          "Stocks"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CancelOrderResponse": {
      "type": "object"
    },
    "v1CreateOrderResponse": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string"
        }
      }
    },
    "v1GetOrderInfoResponse": {
      "type": "object",
      "properties": {
        "status": {
          "$ref": "#/definitions/v1OrderStatus"
        },
        "userId": {
          "type": "string",
          "format": "int64"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Item"
          }
        }
      }
    },
    "v1GetStockInfoResponse": {
      "type": "object",
      "properties": {
        "count": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "v1Item": {
      "type": "object",
      "properties": {
        "sku": {
          "type": "integer",
          "format": "int64"
        },
        "count": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "v1OrderStatus": {
      "type": "string",
      "enum": [
        "ORDER_STATUS_UNSPECIFIED",
        "ORDER_STATUS_NEW",
        "ORDER_STATUS_AWAITING_PAYMENT",
        "ORDER_STATUS_FAILED",
        "ORDER_STATUS_PAYED",
        "ORDER_STATUS_CANCELED"
      ],
      "default": "ORDER_STATUS_UNSPECIFIED"
    },
    "v1PayOrderResponse": {
      "type": "object"
    }
  }
}